# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mavros Connect API - SaaS Hotspot Management Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants (super-admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"

    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/me:
    get:
      operationId: getMyTenant
      tags: [tenants]
      summary: Get current user's tenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  # ─── LOCATIONS ───────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users for tenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a staff user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/suspend:
    post:
      operationId: suspendUser
      tags: [users]
      summary: Suspend a hotspot user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/unsuspend:
    post:
      operationId: unsuspendUser
      tags: [users]
      summary: Unsuspend a hotspot user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unsuspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── ROUTERS ─────────────────────────────────────────────────────────────────
  /routers:
    get:
      operationId: listRouters
      tags: [routers]
      summary: List all routers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of routers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HotspotRouter"

    post:
      operationId: createRouter
      tags: [routers]
      summary: Add a new router
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouterInput"
      responses:
        "201":
          description: Router added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotspotRouter"
  /routers/{id}:
    get:
      operationId: getRouter
      tags: [routers]
      summary: Get router by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Router
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotspotRouter"

    patch:
      operationId: updateRouter
      tags: [routers]
      summary: Update router
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouterUpdate"
      responses:
        "200":
          description: Updated router
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotspotRouter"

    delete:
      operationId: deleteRouter
      tags: [routers]
      summary: Delete router
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /routers/{id}/test:
    post:
      operationId: testRouterConnection
      tags: [routers]
      summary: Test MikroTik API connection
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouterConnectionTest"
  /routers/{id}/active-users:
    get:
      operationId: getRouterActiveUsers
      tags: [routers]
      summary: Get active hotspot users on router
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Active users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActiveUser"

  # ─── PACKAGES ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, tenantId, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, admin, manager, staff]
        tenantId:
          type: ["integer", "null"]
        locationId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, suspended, inactive]
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [email, name, role, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, staff]
        password:
          type: string
          minLength: 8
        locationId:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, staff]
        locationId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, suspended, inactive]

    # ── Tenant ───────────────────────────────────────────────────────────────
    Tenant:
      type: object
      required: [id, name, slug, status, plan, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended, trial]
        plan:
          type: string
          enum: [starter, professional, enterprise]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TenantInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        plan:
          type: string
          enum: [starter, professional, enterprise]
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
        primaryColor:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        status:
          type: string
          enum: [active, suspended, trial]
        plan:
          type: string
          enum: [starter, professional, enterprise]

    # ── Location ─────────────────────────────────────────────────────────────
...
    HotspotRouter:
      type: object
      required: [id, tenantId, locationId, name, ipAddress, apiPort, status, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        locationId:
          type: integer
        name:
          type: string
        ipAddress:
          type: string
        apiPort:
          type: integer
        username:
          type: string
        status:
          type: string
          enum: [online, offline, unknown]
        model:
          type: ["string", "null"]
        firmwareVersion:
          type: ["string", "null"]
        lastSeen:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    RouterInput:
      type: object
      required: [name, ipAddress, apiPort, username, password, locationId]
      properties:
        name:
          type: string
        ipAddress:
          type: string
        apiPort:
          type: integer
        username:
          type: string
        password:
          type: string
        locationId:
          type: integer
        model:
          type: string
    RouterUpdate:
      type: object
      properties:
        name:
          type: string
        ipAddress:
          type: string
        apiPort:
          type: integer
        username:
          type: string
        password:
          type: string
        locationId:
          type: integer
    RouterConnectionTest:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        routerInfo:
          type: ["object", "null"]
          properties:
            model:
              type: string
            version:
              type: string
            uptime:
              type: string
    ActiveUser:
      type: object
      required: [mac, ip, username, uptime, bytesIn, bytesOut]
      properties:
        mac:
          type: string
        ip:
          type: string
        username:
          type: string
        uptime:
          type: string
        bytesIn:
          type: integer
        bytesOut:
          type: integer

    # ── Package ──────────────────────────────────────────────────────────────
...