# lib/api-zod/src/generated/types/hotspotRouter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mavros Connect API - SaaS Hotspot Management Platform
 * OpenAPI spec version: 0.1.0
 */
import type { HotspotRouterStatus } from './hotspotRouterStatus';

export interface HotspotRouter {
  id: number;
  tenantId: number;
  locationId: number;
  name: string;
  ipAddress: string;
  apiPort: number;
  username?: string;
  status: HotspotRouterStatus;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  firmwareVersion?: string | null;
  /** @nullable */
  lastSeen?: Date | null;
  createdAt: Date;
}