# lib/api-zod/src/generated/types/routerConnectionTest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mavros Connect API - SaaS Hotspot Management Platform
 * OpenAPI spec version: 0.1.0
 */
import type { RouterConnectionTestRouterInfo } from './routerConnectionTestRouterInfo';

export interface RouterConnectionTest {
  success: boolean;
  message: string;
  /** @nullable */
  routerInfo?: RouterConnectionTestRouterInfo;
}