# lib/api-zod/src/generated/types/routerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mavros Connect API - SaaS Hotspot Management Platform
 * OpenAPI spec version: 0.1.0
 */

export interface RouterInput {
  name: string;
  ipAddress: string;
  apiPort: number;
  username: string;
  password: string;
  locationId: number;
  model?: string;
}